Move requests in separate files
